requirements.txt


coreapi
coreschema

Django==3.0
django-allauth==0.41.0
django-cors-headers
django-environ
django_filter==2.4.0
django-heroku
django-rest-auth
django-rest-registration
django-rest-swagger
djangorestframework
djangorestframework-filters
django-utils-six
drf-yasg

Gunicorn

Pillow==8.4.0
psycopg2==2.8.6

qrcode
slugify==0.0.1

yarl


